WT-602: Removed Task Management legacy code
