+++ client/src/components/CRM/Tasks/TasksMain.tsx
@@ -1,15 +1,15 @@
-/* eslint-disable react/react-in-jsx-scope */
+import React, {ReactElement} from 'react';
 import Navigation from '../Navigation';
-import TaskManager from './TaskManager';
+import Tasks from './Tasks';
 
 /**
  * @returns {object} TasksMain - has Navigation
  *  and Tasks to reroute to other components
  */
-const TasksMain=() =>{
+const TasksMain = (): ReactElement =>{
   return (
-    <div className='flex flex-row  '>
-      <TaskManager/>
+    <div className='flex flex-row'>
+      <Tasks/>
       <Navigation/>
     </div>
   );
